+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.VkNewsClient"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.mainScreen.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.VkNewsClient">
