+++ app/src/main/java/com/voitov/vknewsclient/presentation/commentsScreen/CommentsScreen.kt
@@ -1,4 +1,4 @@
-package com.voitov.vknewsclient.ui.theme.commentsScreen
+package com.voitov.vknewsclient.presentation.commentsScreen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.padding
@@ -10,10 +10,8 @@ import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.voitov.vknewsclient.ui.theme.VkNewsClientTheme
 
 @Composable
 fun CommentsScreen(
