+++ app/src/main/java/com/voitov/vknewsclient/presentation/mainScreen/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.voitov.vknewsclient
+package com.voitov.vknewsclient.presentation.mainScreen
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,18 +8,18 @@ import androidx.compose.runtime.livedata.observeAsState
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.vk.api.sdk.VK
 import com.vk.api.sdk.auth.VKScope
-import com.voitov.vknewsclient.ui.theme.MainScreen
+import com.voitov.vknewsclient.presentation.mainScreen.MainScreen
 import com.voitov.vknewsclient.ui.theme.VkNewsClientTheme
-import com.voitov.vknewsclient.ui.theme.authorizationScreen.AuthorizationScreen
-import com.voitov.vknewsclient.ui.theme.authorizationScreen.AuthorizationScreenState
-import com.voitov.vknewsclient.ui.theme.authorizationScreen.MainViewModel
+import com.voitov.vknewsclient.presentation.authorizationScreen.AuthorizationScreen
+import com.voitov.vknewsclient.presentation.authorizationScreen.AuthorizationScreenState
+import com.voitov.vknewsclient.presentation.authorizationScreen.AuthorizationViewModel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             VkNewsClientTheme {
-                val viewModel: MainViewModel = viewModel()
+                val viewModel: AuthorizationViewModel = viewModel()
                 val authorizationState = viewModel.authorizationState.observeAsState(
                     AuthorizationScreenState.InitialState
                 )
