+++ app/src/main/java/com/voitov/vknewsclient/presentation/mainScreen/NavigationItem.kt
@@ -1,4 +1,4 @@
-package com.voitov.vknewsclient.ui.theme
+package com.voitov.vknewsclient.presentation.mainScreen
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Favorite
